backend/requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.1
aioschedule==0.5.2
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
et-xmlfile==1.1.0
fastapi==0.104.1
frozenlist==1.4.1
h11==0.14.0
idna==3.4
magic-filter==1.0.12
motor==3.3.1
multidict==6.0.4
numpy==1.26.3
openpyxl==3.1.2
pandas==2.1.4
pydantic==2.4.2
pydantic-settings==2.1.0
pydantic_core==2.10.1
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.23.2
yarl==1.9.4
